fix: dial bootnode for sync
