+++ crates/pathfinder/src/p2p_network.rs
@@ -1,3 +1,4 @@
+pub mod common;
 pub mod consensus;
 pub mod identity;
 pub mod sync;
