+++ crates/pathfinder/src/p2p_network/consensus.rs
@@ -29,11 +29,13 @@ mod inner {
 
     use anyhow::Context;
     use futures::FutureExt;
-    use p2p::libp2p::multiaddr::{Multiaddr, Protocol};
+    use p2p::libp2p::multiaddr::Protocol;
     use pathfinder_common::ChainId;
 
     use super::*;
     use crate::config::p2p::P2PConsensusConfig;
+    use crate::p2p_network::common::dial_bootnodes;
+    use crate::p2p_network::common::ensure_peer_id_in_multiaddr;
     use crate::p2p_network::identity;
 
     #[tracing::instrument(name = "p2p", skip_all)]
@@ -89,47 +91,7 @@ mod inner {
                 .with_context(|| format!("Starting consensus P2P listener: {addr}"))?;
         }
 
-        let ensure_peer_id_in_multiaddr = |addr: &Multiaddr, msg: &'static str| {
-            addr.iter()
-                .find_map(|p| match p {
-                    Protocol::P2p(peer_id) => Some(peer_id),
-                    _ => None,
-                })
-                .ok_or_else(|| anyhow::anyhow!(msg))
-        };
-
-        let mut success = false;
-        for bootstrap_address in bootstrap_addresses {
-            let peer_id = match ensure_peer_id_in_multiaddr(
-                &bootstrap_address,
-                "Bootstrap addresses must include peer ID",
-            ) {
-                Ok(id) => id,
-                Err(e) => {
-                    tracing::warn!("Invalid bootstrap address {bootstrap_address}: {e:#}");
-                    continue;
-                }
-            };
-
-            match core_client.dial(peer_id, bootstrap_address.clone()).await {
-                Ok(_) => {
-                    success = true;
-                }
-                Err(e) => {
-                    tracing::warn!("Failed dialing {bootstrap_address}: {e:#}");
-                    continue;
-                }
-            };
-
-            let relay_listener_address = bootstrap_address.clone().with(Protocol::P2pCircuit);
-            if let Err(e) = core_client
-                .start_listening(relay_listener_address.clone())
-                .await
-            {
-                tracing::warn!("Failed starting relay listener on {relay_listener_address}: {e:#}");
-            }
-        }
-        if !success {
+        if !dial_bootnodes(bootstrap_addresses, &core_client).await {
             anyhow::bail!("Failed to dial any configured bootstrap node")
         }
